### 2025-05-05 #704 @n-kall
priorsense - Prior Diagnostics and Sensitivity Analysis
Submitting Author Name: Noa Kallioinen
Submitting Author Github Handle: @n-kall
Other Package Authors Github handles: (comma separated, delete if none) @topipa, @paul-buerkner, @avehtari
Repository:  https://github.com/n-kall/priorsense
Version submitted: 1.1.1.9000
Submission type: Stats
Badge grade: bronze/silver/gold (select one)

Version accepted: TBD
---

```
Package: priorsense
Title: Prior Diagnostics and Sensitivity Analysis
Version: 1.1.1.9000
Authors@R: c(person("Noa", "Kallioinen", email = "noa.kallioinen@aalto.fi", role = c("aut", "cre", "cph")),
	     person("Topi", "Paananen", role = c("aut")),
	     person("Paul-Christian", "Bürkner", role = c("aut")),
	     person("Aki", "Vehtari", role = c("aut")),
	     person("Frank", "Weber", role = c("ctb"))
	     )
Description: Provides functions for prior and likelihood sensitivity analysis in Bayesian models. Currently it implements methods to determine the sensitivity of the posterior to power-scaling perturbations of the prior and likelihood.
Roxygen: list(markdown = TRUE, roclets = c ("namespace", "rd", "srr::srr_stats_roclet"))
    checkmate (>= 2.3.1),
    ggdist (>= 3.3.2),
    ggh4x (>= 0.2.5),
    ggplot2 (>= 3.5.1),
    grDevices (>= 3.6.2),
    matrixStats (>= 1.3.0),
    posterior (>= 1.6.0),
    rlang (>= 1.1.4),
    tibble (>= 3.2.1),
    utils
    bayesplot (>= 1.11.1),
    brms (>= 2.22.0),
    cmdstanr (>= 0.8.1),
    iwmm (>= 0.0.1),
    philentropy (>= 0.8.0),
    quarto (>= 1.4.4),
    R2jags (>= 0.8),
    rstan (>= 2.32.6),
    transport (>= 0.15),
    vdiffr (>= 1.0.8)
VignetteBuilder: quarto
Additional_repositories:
    https://topipa.r-universe.dev,
    https://stan-dev.r-universe.dev
Config/Needs/website: quarto
```

- Please indicate which of our [statistical package categories](https://stats-devguide.ropensci.org/overview.html#overview-categories) this package falls under. (Please check one or more appropriate boxes below):

- [x] A pre-submission inquiry has been approved in [issue#697](#697)

The target audience is users of Bayesian models who would like to check the prior (and likelihood) sensitivity of their model. This can demonstrate the robustness of the results, as well as how important the prior choice is.

    - *The first implementation of a novel algorithm*; or
    - *The first implementation within* **R** *of an algorithm which has previously been implemented in other languages or contexts*; or
    - *An improvement on other implementations of similar algorithms in* **R**.

Response:

The \pkg{priorsense} package provides functions for prior and likelihood sensitivity analysis of Bayesian models. Currently it implements methods to determine the sensitivity of the posterior to power-scaling perturbations of the prior and likelihood and is the first implementation of the method described in Kallioinen et al. (2023).

Response: Not applicable

Gold

- Compliance with a good number of standards  beyond those identified as minimally necessary
- Have a demonstrated generality of usage beyond one single envisioned use case: While originally developed for Stan models, priorsense has been extended to work with brms models, JAGS models and arbitrary posterior draws.
- Demonstrating excellence in compliance with multiple standards from at least two broad sub-categories: Aiming for excellence in documentation (6.1.1 and 6.3.1) with examples and vignettes, and (6.3.5) Visualization and summarization output

Response:
priorsense is already on CRAN
