### 2025-07-03 0a1b2c6
updates for new ggplot2 version
