+++ DESCRIPTION
@@ -1,6 +1,6 @@
 Package: priorsense
 Title: Prior Diagnostics and Sensitivity Analysis
-Version: 1.1.1.9000
+Version: 1.1.2
 Authors@R: c(person("Noa", "Kallioinen", email = "noa.kallioinen@aalto.fi", role = c("aut", "cre", "cph")),
 	     person("Topi", "Paananen", role = c("aut")),
 	     person("Paul-Christian", "Bürkner", role = c("aut")),
@@ -17,7 +17,7 @@ Imports:
     checkmate (>= 2.3.1),
     ggdist (>= 3.3.2),
     ggh4x (>= 0.2.5),
-    ggplot2 (>= 3.5.1),
+    ggplot2 (>= 3.5.2.9001),
     grDevices (>= 3.6.2),
     matrixStats (>= 1.3.0),
     posterior (>= 1.6.0),
