### 2025-07-03 0985798
add plot method
