+++ NAMESPACE
@@ -13,6 +13,7 @@ S3method(log_lik_draws,stanfit)
 S3method(log_prior_draws,CmdStanFit)
 S3method(log_prior_draws,draws)
 S3method(log_prior_draws,stanfit)
+S3method(plot,powerscale_sequence)
 S3method(plot,priorsense_plot)
 S3method(posterior::summarise_draws,powerscaled_draws)
 S3method(posterior::summarise_draws,powerscaled_sequence)
