+++ DESCRIPTION
@@ -16,8 +16,8 @@ RoxygenNote: 7.3.2
 Imports:
     checkmate (>= 2.3.1),
     ggdist (>= 3.3.2),
-    ggh4x (>= 0.2.5),
-    ggplot2 (> 3.5.2),
+    ggh4x (>= 0.3.1),
+    ggplot2 (>= 4.0.0),
     grDevices (>= 3.6.2),
     matrixStats (>= 1.3.0),
     posterior (>= 1.6.0),
