### 2025-10-17 bacfe0c
fix issue with plotting high pareto-k in quantities plot
