### 2025-10-23 a6cd49e
handle `variables` as alias of `variable`, use ggplot2 ecdf
