+++ R/plots.R
@@ -8,6 +8,8 @@
 ##' @template plot_args
 ##' @template div_measure_arg
 ##' @template resample_arg
+##' @param type Character specifying type of plot, one of "dens",
+##'   "ecdf",or "quantities".
 ##' @template ggplot_return
 ##' @section Plot Descriptions: \describe{
 ##'   \item{`powerscale_plot_dens()`}{ Kernel density plot of
@@ -18,7 +20,7 @@
 ##'   \item{`powerscale_plot_quantities()`}{ Plot of posterior
 ##'   quantities with respect to power-scaling.} }
 ##'
-##' @srrstats {EA5.0} 
+##' @srrstats {EA5.0}
 ##' @srrstats {EA5.0a} default typefaces in ggplot2 are accessible
 ##' @srrstats {EA5.0b} default color scheme chosen to be accessible
 ##' @srrstats {EA5.4} values are rounded before plotting
@@ -34,6 +36,7 @@ NULL
 ##' prepare plot data
 ##' @param x priorsense_data object
 ##' @param variable character vector specifying variables to plot
+##' @param variables alias of `variable`.
 ##' @param resample boolean flag specifying whether to resample
 ##' @param ... unused
 ##' @return data frame with data for plotting
@@ -106,11 +109,11 @@ prepare_plot_data <- function(x, variable, resample, ...) {
 
       likelihood_draws[[i]][[".powerscale_alpha"]] <-
         likelihood_ps_details$alpha
-      
+
       likelihood_draws[[i]]$component <- "likelihood"
-      
+
       likelihood_draws[[i]]$pareto_k <- likelihood_ps_details$diagnostics$khat
-      
+
       likelihood_draws[[i]]$pareto_k_threshold <-
         likelihood_ps_details$diagnostics$khat_threshold
 
@@ -251,6 +254,7 @@ powerscale_plot_dens <- function(x, ...) {
 powerscale_plot_dens.default <-
   function(x,
            variable = NULL,
+           variables = NULL,
            length = 3,
            resample = FALSE,
            intervals = c(0.5, 0.8, 0.95),
@@ -265,6 +269,7 @@ powerscale_plot_dens.default <-
     powerscale_plot_dens(
       ps,
       variable = variable,
+      variables = variables,
       length = length,
       resample = resample,
       intervals = intervals,
@@ -281,6 +286,7 @@ powerscale_plot_dens.default <-
 powerscale_plot_dens.powerscaled_sequence <-
   function(x,
            variable = NULL,
+           variables = NULL,
            resample = FALSE,
            intervals = c(0.5, 0.8, 0.95),
            trim = NULL,
@@ -294,6 +300,16 @@ powerscale_plot_dens.powerscaled_sequence <-
            ) {
 
     # input checks
+      if (!is.null(variable) && !is.null(variables)) {
+   checkmate::assert(
+      if (identical(variable, variables)) TRUE else "must be identical if both provided",
+      .var.name = "`variable` and `variables`"
+      )
+  }
+  if (is.null(variable)) {
+    variable <- variables
+  }
+
     checkmate::assert_character(variable, null.ok = TRUE)
     checkmate::assert_logical(resample, len = 1)
     checkmate::assert_logical(help_text, len = 1)
@@ -505,6 +521,7 @@ powerscale_plot_ecdf <- function(x, ...) {
 powerscale_plot_ecdf.default <-
   function(x,
            variable = NULL,
+           variables = NULL,
            length = 3,
            resample = FALSE,
            facet_rows = "component",
@@ -519,6 +536,7 @@ powerscale_plot_ecdf.default <-
     powerscale_plot_ecdf(
       ps,
       variable = variable,
+      variables = variables,
[…]
