+++ R/print.R
@@ -117,3 +117,46 @@ print.whitened_draws_summary <- function(x, ...) {
   invisible(x)
 
 }
+
+
+##' @export
+print.exptilting_details <- function(x, ...) {
+
+  pareto_k <- x$diagnostics$khat
+  pareto_k_threshold <- x$diagnostics$khat_threshold
+  pareto_kf <- x$diagnostics$khatf
+
+  pareto_k_print <- c()
+
+  pareto_k_print <- paste("pareto-k:", round(pareto_k, digits = 2), "\n")
+
+  if (!is.null(pareto_kf)) {
+    pareto_k_print <- c(
+      "moment-matched\n",
+      pareto_k_print,
+      paste("pareto-kf",
+            round(pareto_kf, digits = 2),
+            "\n")
+    )
+  }
+
+  cat(
+    "\nexponential tilting\n",
+    paste("eta:", x$eta, "\n"),
+    paste("tilted variable:", x$tilted_variable, "\n"),
+    pareto_k_print,
+    paste("pareto-k threshold:", round(pareto_k_threshold, 2), "\n"),
+    paste("resampled:", x$resampled, "\n"),
+    paste("transform:", x$transform_details$transform, "\n")
+  )
+
+  invisible(x)
+}
+
+##' @export
+print.exptilted_draws <- function(x, ...) {
+  NextMethod(...)
+  print(attr(x, "exptilting"), ...)
+
+  invisible(x)
+}
