+++ R/exptilt.R
@@ -1,0 +1,298 @@
+##' Exponential tilting perturbation
+##'
+##' Estimate posterior draws based on exponential perturbations of
+##' prior using importance sampling (and optionally moment matching).
+##'
+##' @name exptilt-overview
+##'
+##' @template fit_arg
+##' @param variable Vector of variable names to return estimated
+##'   posterior draws for. If `NULL` all variables will be included.
+##' @param variables Alias of `variable`.
+##' @template selection_arg
+##' @template prediction_arg
+##' @param ... Further arguments passed to internal functions.
+##' @return An `exptilted_draws` or `exptilted_sequence` object,
+##'   which contains the estimated posterior draws resulting from the
+##'   power-scaling perturbations and details of the perturbation and
+##'   estimation methods.
+##' @template powerscale_references
+##' @srrstats {G2.0} Assertions are made on the lengths of inputs via the checkmate package
+##' @srrstats {G2.1} Assertions on types of inputs are made via the checkmate package
+##' @examples
+##' ex <- example_powerscale_model()
+##'
+##' exptilt(ex$draws, tilted_variable = "sigma", eta = 0.5)
+##'
+##' @export
+exptilt <- function(x, ...) {
+  UseMethod("exptilt")
+}
+
+
+##' @rdname exptilt-overview
+##' @export
+exptilt.default <- function(x, tilted_variable, eta, tilt_fun = I,
+                            moment_match = FALSE,
+                            k_threshold = NULL,
+                            resample = FALSE,
+                            transform = NULL,
+                            prediction = NULL,
+                            variable = NULL,
+                            variables = NULL,
+                            tilt_fun_args = NULL,
+                            ...) {
+
+  psd <- create_priorsense_data(
+    x,
+    ...
+  )
+  exptilt(
+    psd,
+    tilted_variable = tilted_variable,
+    eta = eta,
+    tilt_fun = tilt_fun,
+    moment_match = moment_match,
+    k_threshold = k_threshold,
+    resample = resample,
+    transform = transform,
+    prediction = prediction,
+    variable = variable,
+    variables = variables,
+    tilt_fun_args = tilt_fun_args
+  )
+}
+
+
+##' @rdname exptilt-overview
+##' @srrstats {G2.3}
+##' @srrstats {G2.3a} checkmate functions used to allow only specific
+##'   argument values
+##' @srrstats {G2.3b} tolower() used for component
+##' @srrstats {G2.4} Input coercion
+##' @srrstats {G2.4b} Input coercion with `as.numeric()`
+##' @srrstats {G2.4c} Input coercion with `as.character()`
+##' @srrstats {G2.4e} Input coercion
+##' @srrstats {EA2.6} vector inputs are coerced to numeric
+##' @export
+exptilt.priorsense_data <- function(x,
+                                    tilted_variable,
+                                    eta,
+                                    tilt_fun = I,
+                                    moment_match = FALSE,
+                                    k_threshold = NULL,
+                                    resample = FALSE,
+                                    transform = NULL,
+                                    prediction = NULL,
+                                    variable = NULL,
+                                    variables = NULL,
+                                    tilt_fun_args = NULL,
+                                    ...) {
+
+  # input coercion
+  eta <- as.numeric(eta)
+  moment_match <- as.logical(moment_match)
+  if (!is.null(k_threshold)) {
+    k_threshold <- as.numeric(k_threshold)
+  }
[…]
