+++ NAMESPACE
@@ -5,8 +5,6 @@ S3method(create_priorsense_data,default)
 S3method(create_priorsense_data,draws)
 S3method(create_priorsense_data,rjags)
 S3method(create_priorsense_data,stanfit)
-S3method(exptilt,default)
-S3method(exptilt,priorsense_data)
 S3method(find_alpha_threshold,default)
 S3method(find_alpha_threshold,priorsense_data)
 S3method(log_lik_draws,CmdStanFit)
@@ -50,7 +48,6 @@ S3method(summarise_draws,whitened_draws)
 export(cjs_dist)
 export(create_priorsense_data)
 export(example_powerscale_model)
-export(exptilt)
 export(log_lik_draws)
 export(log_prior_draws)
 export(powerscale)
