### 2025-10-23 d4e14da
bump brms suggested version
