### 2025-10-23 7845656
cleanup summarise_draws
