### 2025-10-23 a640538
add summarise_draws method for exptilt
