+++ R/summarise_draws.R
@@ -241,3 +241,80 @@ summarise_draws.whitened_draws <- function(.x, ...) {
   class(summary) <- c("whitened_draws_summary", class(summary))
   return(summary)
 }
+
+##' @export
+summarise_draws.exptilted_draws <- function(.x,
+                                            ...,
+                                            .num_args = NULL,
+                                            .args = list(),
+                                            base_draws = NULL,
+                                            diagnostics = FALSE,
+                                            div_measures = "cjs_dist",
+                                            measure_args = list(),
+                                            resample = FALSE) {
+
+  funs <- c(...)
+  if (length(funs) == 0) {
+    funs <- posterior::default_summary_measures()
+  }
+
+  et_details <- get_exptilting_details(.x)
+
+  .args <- as.list(.args)
+
+  if (resample && !et_details$resampled) {
+    # only resample if specified and draws are not already resampled
+    target_draws <- posterior::resample_draws(
+      posterior::merge_chains(.x)
+    )
+  } else {
+    target_draws <- .x
+  }
+
+  if (!resample && !et_details$resampled) {
+    # without resampling, only weighted quantities used
+    funs <- as.list(
+      paste0(funs, "_weighted")
+    )
+
+    # add the weights to args
+    .args <- c(
+      list(weights = stats::weights(.x)),
+      .args
+    )
+  }
+
+  class(target_draws) <- class(target_draws)[-1]
+
+  summ <- posterior::summarise_draws(
+    .x = target_draws,
+    ... = funs,
+    .args = .args,
+    .num_args = .num_args
+  )
+
+
+  if (!is.null(base_draws)) {
+    # calculate the divergences between the base and target draws
+    divergences <- measure_divergence(
+      draws1 = posterior::merge_chains(base_draws),
+      draws2 = posterior::merge_chains(target_draws),
+      measure = div_measures,
+      measure_args = measure_args
+    )
+
+    summ <- merge(summ, divergences, by = "variable")
+  }
+
+  out <- summ
+
+  if (resample) {
+    et_details$resampled <- TRUE
+  }
+
+  attr(out, "exptilting") <- et_details
+
+
+  class(out) <- c("exptilted_draws_summary", class(out))
+  return(out)
+}
