+++ NAMESPACE
@@ -43,7 +43,6 @@ S3method(print,powerscaling_details)
 S3method(print,priorsense_plot)
 S3method(print,whitened_draws)
 S3method(print,whitened_draws_summary)
-S3method(summarise_draws,exptilted_draws)
 S3method(summarise_draws,whitened_draws)
 export(cjs_dist)
 export(create_priorsense_data)
