### 2025-10-23 f4c0074
cleanup print
