### 2025-10-23 ab690ec
add print method
