### 2025-10-23 a4e0b8d
cleanup namespace
