+++ NAMESPACE
@@ -33,8 +33,6 @@ S3method(powerscale_sensitivity,priorsense_data)
 S3method(powerscale_sensitivity,stanfit)
 S3method(powerscale_sequence,default)
 S3method(powerscale_sequence,priorsense_data)
-S3method(print,exptilted_draws)
-S3method(print,exptilting_details)
 S3method(print,powerscaled_draws)
 S3method(print,powerscaled_draws_summary)
 S3method(print,powerscaled_sensitivity_summary)
