### 2025-10-23 230008c
Development (#115)
* fix issue with plotting high pareto-k in quantities plot

* add snapshots for plots

* update cran submission

* reference update (#113)

* Update R-CMD-check.yaml

* Update README.md

* handle `variables` as alias of `variable`, use ggplot2 ecdf

* bump brms suggested version
