### 2026-05-03 #704 @maurolepore
I see both reviews are in (thanks @DylanDijk and @staylo16). We look forward to your response.

Apologies — they were never formally logged with the bot, which is why @staylo16 received an auto-reminder after submitting the review. Logging them now so the state catches up.

Thanks for your patience and for sharing your work with the rOpenSci community.
