### 2026-05-21 #704 @n-kall
## Response to reviewers for priorsense

Thank you @staylo16 and @DylanDijk for your valuable comments on the package. And thank you to all editors who have helped handle this submission. It has been very helpful to get external eyes on the software and in particular the documentation. I believe that all comments and suggestions have been incorporated into the package, and that now the package is greatly improved (especially the documentation). The latest github version on the `main` branch has all the changes, as does the webpage https://n-kall.github.io/priorsense. I will hold off submitting to CRAN for some days, in case there are some further comments.

Below are responses to each specific comment:

## Reviewer 1 (@staylo16)

### Updates to documentation, example code
- The README, webpage, and vignettes have been overhauled. There are now separate vignettes explaining how to adapt model code for each of the supported modelling languages
- All model code examples has been updated to no longer have `prior_alpha` or `likelihood_alpha` in the code, to avoid confusion, as it is not needed
- Identified typos have been fixed
- Documentation on the `div_measure` has been updated to point to each external function
- `Details` section for `powerscale-overview` has been added to give a description of the method

### Updates to code
- plot functions now accept both `colors` and `colours`
- the specified length of the alpha sequence is now consistent with the
  output. In the case of `symmetric = TRUE`, the documentation states that this
  will be rounded down to an odd number (because in order for the sequence to be
  symmetric it must be odd, as it always contains alpha = 1)
- warnings about bandwidth selection are no longer printed
- I cannot reproduce the `length = 2` warning any more

## Reviewer 2 (@DylanDijk)
- the maintainer field is indeed handled by `Authors@R`, and is confirmed to be correct
- The README, webpage, and vignettes have been overhauled. There are now separate vignettes explaining how to adapt model code for each of the supported modelling languages
- a more complex workflow example (airquality) is available in an online-only vignette, and is now linked from the README
