### 2026-08-28 #704 @maelle
@n-kall I see the repository hasn't been transferred. Do you want to transfer it to the rOpenSci organization or would you like us to register your package without your transferring it? Thank you, and congrats on the approval!
