### 2026-08-31 #704 @n-kall
Thank you @maelle for following this up. Sorry for the delay: I had been organising some details regarding this. I would now like to opt-out of transferring the repository to the ROpenSci.

The piorsense repository will be transferred shortly to the @stan-dev organisation. Would it make sense to complete that transfer before registering the package (so that links etc. won't need to be changed)?
